Extract conversion settings from DateTimeZoneProviders
Fixes #1505.

I've left a deprecated property for DateTimeZoneProviders.Serialization.
While moving to 3.0 means we *can* break anything, we actually
restricted the breaking changes to "removing binary serialization"
(and removing incorrectly named method that's been deprecated since
2.1 was released in 2017).

The ForTypeConverter property is new in 3.0, so doesn't need
deprecation. I'll add it to release notes for any existing beta
users, but it should be okay to move. (I'll release a new beta
shortly after this goes in.)